move some classes from pf2e to their own files
